Renamed base-modal path
